update src.vex.units_common docs
